separating helpers 1
